+++ dwtools/TableOfReal_and_SVD.h
@@ -2,7 +2,7 @@
 #define _TableOfReal_and_SVD_h_
 /* TableOfReal_and_SVD.h
  *
- * Copyright (C) 1993-2011, 2015 David Weenink
+ * Copyright (C) 1993-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,7 +30,7 @@ autoSVD TableOfReal_to_SVD (TableOfReal me);
 
 autoGSVD TablesOfReal_to_GSVD (TableOfReal me, TableOfReal thee);
 
-autoTableOfReal SVD_to_TableOfReal (SVD me, long from, long to);
+autoTableOfReal SVD_to_TableOfReal (SVD me, integer from, integer to);
 
 autoTableOfReal SVD_extractLeftSingularVectors (SVD me);
